localized fonts and finished conditional formatting
formatting is super hard-coded, like mega hardcoded, as in i would have to rewrite it from the ground up if i wanted to add more formats. I already had to rewrite it to add in-line bold text, but whatever. Also localized fonts in an attempt to eventually have a .exe that runs